Merge RabbitMQ and gRPC into dev

+++ src/cameras/configurations.py
@@ -11,7 +11,7 @@ class Configurations:
         self.recording = configurations["recording"]
 
     @property
-    def sensitivity(self) -> int:
+    def sensitivity(self) -> float:
         return self._sensitivity
 
     @sensitivity.setter
